Created Components module
Started the code cleanup of the spaghetti code.

Created a 'components' module that contains all the functions that return HTML markup. I didn't really know what else to call it so this will have to be good enough.

Lots of work to do as far as cleaning up and documenting my code... oops?